Add simple tests for DeprecationWarning behaviors
This is a little bit to establish a space for a testsuite, but also a simple
set of three tests to make sure we're warning at instantiation time, and not
some other time (like import time, which could be a mess).
So far, these are pretty weaksauce tests, but they're better than nothing.
Starts to define parameters for a testsuite.