Allow config values to be loaded from environment
Also sets 'GLOBUS_AUTH_TOKEN' to be the first and only value which has this
behavior.
This behavior is optional, but now available for all config values. A config
value's name, uppercased and prefixed with 'GLOBUS_' is looked for in
os.environ, and the result supersedes any values found by loading config. A
worthwhile note is that you can compare this with the value found by loading
the same key with the env check turned off, so we could do nice messaging if
we're worried that people are not understanding the behavior.
I think higher precedence for env variables is the expected case, since it
allows you to set an environment variable in a script to explicitly override
the "usual" behavior of a command.