Refactor config parsing to be simpler and safer
The GlobusConfigParser is now a direct descendant of `object`, not a
`SafeConfigParser` or other descendant of `ConfigParser`. Instead of being a
config parser, it contains a config parser. That lets us define a good `read()`
in terms of the "Public API" `read()` method of ConfigParsers, and not try to
interact with `_read()` or other more fragile elements of the ConfigParser
package. (Not fragile in the sense that they don't work well or break easily,
but in the sense that depending on their exact behaviors is fragile.)
Additionally, loading config from file, looking up keys, with slightly extended
logic, and knowing the name of the [general] config section are all now
responsibilities of the GlobusConfigParser class.

The rest of the module, as a result, gets to go a lot more barebones. Tightens
up the singleton at the module level in a clearer-cut function whose sole
purpose is to instantiate the one and only instance of GlobusConfigParser.
The only functions which are "public" parts of the module are still
'get_auth_token()' and 'get_service_url', which still have the same signatures.