Add more args to TransferClient.endpoint_search
These make the helper more meaningful/useful to look at as a description
of what to do, since they provide more guidance as to what the desired
arguments are for an endpoint search.
Additionally, adds a docstring that includes some simple examples of
endpoint searches.
Adds a helper that merges together given arguments and the catchall
**kwargs to produce a set of parameters for `BaseClient.get()`