Attempt more generic and safer paginated EP search
Generalized pagination in the Transfer API to a decorator that wraps up
a method to be a paginated version of itself. Documents the semantics of
the decorator, assumptions that it makes, and recommended usage.
Supports a paginated=True/False kwarg to paginated funcs with a default
of "False" so that naive SDK users won't be requesting large, paged
result sets.
Lots of this is just doc for clarity on decorator semantics because
they're weeeeeeiiird.
Adds a PaginationOverrunError exception type, which inherits directly
from GlobusError, as the type of error raised if you request more
results than the API will allow.

endpoint_search is now decorated with the new @paginated_resource
decorator, in all of its gory glory.
Still needs docs for how to invoke it.