Documentation expansion for paginated search
Includes a small number of bugfixes to pagination, but the main body of
work here is a large expansion to the docstring of endpoint_search to
cover the paginated version of it.
Describes how to use `paginate=True` and `limit=N`, but omits any
mention of offset. It's for advanced API users only, and they'll know
about it if they read the API docs.