Tidy up PaginatedResource decorator significantly
It's now a class style decorator, which actually makes the pattern
significantly more readable. Trimmed unnecessary and/or confusing
elements of the documentation, and many of the magical parameters of the
decorator are now clearly parts of its initializer and bound to an
instance of the decorator callable.
Also dropped some unnecessary arguments to the decorator to make things
simpler.