Make response types uniform
All return values from client calls (excepting Nexus) are now
GlobusRepsonse objects of one form or another. To distinguish, the
preexisting GlobusResponse object has been renamed to be a
GlobusHTTPResponse, identifying it as the result of some HTTP call. The
response data is still stored the same way, but "_underlying_response"
is now just "data", as part of the now typical GlobusResponse structure.
Defines str and repr identically for GlobusResponse, showing repr() of
the underlying "data" object in each case.