Wrap GlobusResponse.data in a property
Also a couple of fixes for flake8.
Rather than letting callers directly access the data attribute, wrapping
it in a property gives subclasses a nice place to override the meaning
of that property. Importantly, in the case of a GlobusHTTPResponse, it
can just wrap the json_body method.
After a discussion with @jaswilli of the possible behavior here, I
decided to go with `None` for cases in which decoding fails. In a
certain sense, that's a response having "no data" of which to speak. But
you can of course still inspect the HTTP status. That lines up nicely
with the typical way of handling responses like 5xx or 404, in which the
body really doesn't matter and you only care about the status code.
If a caller really wants the raw text body for some reason, it's still
available.