Add transfer submission helpers
Uses the pattern of make_<methodname>_* to clarify that these are
helpers associated strongly with the submit_transfer method. It's a bit
verbose, but `make_submit_transfer_data` is unlikely to collide with
other functionality in the future and cause confusion.
Includes handling of the submission ID, since this is a logical place
for it.