Handle per-service tokens
Now that tokens.globus.org is a thing (for now, as a stopgap measure),
we can move off of NexusClient. Because of a mismatch in our
understanding of how this was going to work, some of our prior naming
was not so great. A little bit of name shuffling was necessary to
accomodate.
Makes BaseClient more abstract by putting a `NotImplemented` method into
its initializer -- so it won't be possible to instantiate a straight
BaseClient anymore (not that you'd really want to). If you want a
client to an API that we don't have baked into the SDK, you have to
subclass BaseClient and define a method for loading its token as a string.
Still extensible, but less monkey-patch-able.