Add a custom User Agent for the SDK
Resolves #17
This is a User-Agent header used to identify the SDK to Globus Servers
(for logging, not logical dispatch). It is modifiable via a parameter to
the client constructors to append a "/<app_name>" where "<app_name>" is
a user-supplied string.
Includes SDK version and is nicely compact. Current format looks like
  globus-sdk-py-0.1.0
or
  globus-sdk-py-0.1.0/My Application

Nicely leaves us room to expand with other SDKs in the future, like
  globus-sdk-php-0.0.1/Theoretical PHP App

And should lead to fairly straightforward server-side grepping.