Transfer paging with eager error triggering
The idea here is to force any paged calls to fetch the first page of
results immediately, when the iterable object is instantiated. That way,
when you make a paged call, you get an error at the invocation site, not
at the iteration site.
Protects us partially against people doing something like

```
tc = TransferClient()
try:
    eps = tc.endpoint_search(...)
except GlobusAPIError as e:
    ...

for ep in eps:
    ...
```

Importantly, we behave *much* better with structures like this:

```
try:
    eps = tc.endpoint_search(...)
except GlobusAPIError as e:
    ...
else:
    print('Successfully got endpoint search results!')
    ... # some code that assumes we'll be able to iterate
    for ep in eps:
        ...
```

During iteration, a call to the API could still fail, but a variety of
cases, like a malformed call to endpoint_search, will be caught
immediately. Excepting credential expiration during iteration on a paged
call, it's unlikely that the first call will succeed, but later calls
will fail.
To do this, the `@PaginatedResource` decorator had to go -- binding of
arguments at call time vs. definition time and other issues abounded.
Instead, we keep that name, but `PaginatedResource` is now a class,
instantiated with a function or instance method to wrap, which
implements a custom __iter__ to have an iterable interface. This has the
added benefit of enhancing code clarity, since the return type of paged
functions when decorated no longer mismatches with the return type of
the undecorated functions.

Additionally, this commit enforces those call semantics on the non-paged
API calls by wrapping their results in list comprehensions, rather than
using the `yield` generator function structure.
It's not more or less clean looking, but importantly it doesn't defer
the API call execution until iteration time.