Tidy up PaginatedResource initializer
Mostly, this is a matter of naming. PaginatedResource is not a purely
generic class, so there's no reason for it to use highly ambiguous names
like "func" for a function that it knows *must be* a method of
TransferClient. Fix up the names to be clearer.
On a related note, don't take an argument list when we only ever pass a
single positional argument, the path.