Change presentation of clients in docs
A few things: prioritizes presentation of the high-level interface over
the low-level interface. We don't really want people going to the
low-level interface unless they know exactly what they're doing.
Some of this is reordering, and some is done in content changes.
Also does a simple description and example at the top of the "Developer
Interface" docs, using the high level interface.
Drops all mention of the low-level interface from the top-level index
page.

Expands doc on BaseClient, and adds ":member-order: bysource" to a bunch
of classes.