Add `__getitem__` to GlobusResponses
All GlobusResponse objects have a `__getitem__` method which passes
through to `__getitem__` on the `data` property. If a TypeError is
raised by the attempted lookup (and I was careful to not catch a
TypeError raised by getting `data`), we catch it and reraise as a mildly
specialized TypeError. Instead of seeing that some internal object
doesn't support indexing, we show that whatever kind of GlobusResponse
is being used doesn't support it.
Open question: is that the right way to handle a failure to __getitem__
on the response data?