Iterable responses w/ new response_class feature
The previous pass was very clunky. This is a redesign in which we add a
new feature, much like the behavior of the logging and argparse
libraries, and similar python libs, and allow a class to be passed as a
kwarg in order to customize behavior.
All request methods on BaseClient now accept a `reponse_class` argument,
and the `response_class` attribute on client objects is renamed to
`default_response_class`. When a call is made with a custom response
class, it takes precedence over the `defualt_response_class`, of course.
The result is that iterable `TransferResponse`s are a class of their own,
`IterableTransferResponse`. That lets us define `__iter__` where it
matters, but not have to define a version that explicitly raises
TypeError on classes where it's not appropriate to treat iterability as
a general concept.