Add globus_sdk.authorizers -- generic auth tooling
The idea of a GlobusAuthorizer is that it generates `Authorization`
headers for use with Globus services. Because of the variety of methods
of Authorizing Globus API calls, we need to support, at minimum, Basic
Auth and Bearer Token authorization methods. This is part of a refactor
that will allow us to handle refresh tokens more fluidly moving forward.
The basic idea is this:
- every time BaseClient._request() is called, it looks for an Authorizer
  which can generate an Authorization header
- if the API returns a 401, the Authorizer is asked if it can handle
  401s and try a new Authorization header (e.x. refresh the token being
  used)
- if the Authorizer says that it can handle a 401, and the caller has
  not passed retry_401=False , then the call is tried again *once*

For our current auth techniques, the retry logic is harmless but not
immediately useful. Moving forward, it may be extremely useful.