Remove ability for clients to take creds directly
The AuthClient and TransferClient classes had the ability to consume
their credentials and implicitly create the underlying GlobusAuthorizer
objects to manage their connection with the services. However, this
pattern results in laborious error-checking and doesn't provide a clear
"one way to do it". Ditch it in favor of always specifying the
authorizer explicitly, and rely on a default behavior that replicates
the old default: load a token from config and create an AT authorizer
that uses it.
Get rid of the exclusive params helper, as it is no longer needed.