Significant improvements to documentation
Perhaps most importantly, add examples to the authorizer doc page.
Additionally, rename the GlobusOAuthTokenResponse to just
OAuthTokenResponse -- it reads cleaner in a bunch of places that way.
Additionally, enhance the docstrings for the AuthClient to be more talky
and "explain" the various things that you can do with them.
Reorganizes the examples in the authorization section to come before the
class documentation (which is arguably not even very useful to outside
developers most of the time).