Tweak top-level Client doc to not use Authorizer
The "API" Client doc should not be using an explicit AT Authorizer. This
change was made temporarily, when it looked like we might remove support
for ATs in the config during this set of changes. Since I reverted that,
it's no longer necessary to do explicit auth in the basic doc of
TransferClient.