Replace AuthClient.client_type with subclasses
Rather than having a single class with multivariate behaviors based on
an enumerable argument to its constructor, describe the various types of
Auth Clients as distinct classes.
NativeAppAuthClient and ConfidentialAppAuthClient are the two
specific new classes, but it's still allowable to instantiate an
AuthClient directly.
Importantly, these let us remove dispatch behavior on type and instead
leverage basic classmethod overrides. The subclasses can also do more
sophisticated constructors tailored to their needs.
In order to handle the new documentation for these classes, refactor
docs to put AuthClient (and subclasses) on a separate page from
TransferClient.

To support NativeAppAuthClient, add the NullAuthorizer, which we've
suspected we would need for a while.