Remove temporary options for token management
In two locations, there were parameters to methods --
**additional_params for `AuthClient.oauth2_refresh_token()` and
`no_auth_header` for `BaseClient._request()` and its wrappers -- which
were only necessary for intermediate solutions which we ultimately are
not using. In the presence of the NullAuthorizers, the 'no_auth_header'
arg would be better supplanted by accepting a general override on the
authorizer for a call, but it's not clear that we need that capability.
Similarly, the additional_params kwarg made sense before we
differentiated between AuthClient types using subclasses. With this
newer model, the notion of needing to customize a call on the base class
fades away, and it's not obvious that we'll ever have a case of an
AuthClient descendant which needs that kind of caller-oriented
customization. Customizing calls at the class level has clearer
semantics in terms of the association of a collection of customized
calls.