Initial version of 3-legged OAuth Grant flow
This is the first version of support for the Authorization Code grant,
often casually referred to as 3-legged OAuth. This is a flow manager
like the NativeAppFlowManager which hangs off of the
ConfidentialAppAuthClient. Gave it a wrapping helper, and ensure that
the flow manager is visible in the top-level package namespace.