Add python-jose dependent ID Token parsing
If you have `python-jose` installed, we enable parsing of OIDC ID
Tokens in the SDK. This is needed for applications which want to make
sense of the OIDC information, but because it depends upon pycrypto, it
carries a lot of baggage. For basic SDK usage, we shouldn't be requiring
these libs, which can easily conflict with other existing package sets
and hurt the ease of integration.

Adds a page to docs about optional dependencies, and a new exception
type specifically for handling those dependencies.