Add support Dependent Token Grant
Add support for the Dependent Token Grant to
ConfidentialAppAuthClients. Like the client credentials grant, this is a
single call to the token endpoint, properly formatted and protected with
the client's credentials.

Dependent token grant provides a totally different response from
the other calls to the token endpoint, and requires customizations
around its parsing/interpretation.
Luckily, we can make the response feel samey by giving it a
`by_resource_server` property, indexed in a very similar way.

Also, fix token introspect, which  shouldn't be on the base auth client
-- you need credentials to do it, so move it down. Additionally, token
introspect was wrongly using urlencode, which resulted in bad responses.