Update all docstrings to our new 0.4.0 style
Any instances of the :param: list for sphinx have been replaced with a
"Parameters" section of the docstring, examples are under and "Examples"
section, and links to API docs should be under "External Documentation".
There may be a couple of minor misses here and there, but this covers
the vast majority of the SDK. At least, there are no *known*
non-compliant docstrings.