Add library logging, starting with BaseClient
Configures the top-level logger in globus_sdk.__init__, and adds a
logger whose path is the fully qualified class name of the client inside
of BaseClient. As a result, any client can call look at `self.logger` as
a safe, differentiated logger. It actually will pass their logs through
a LoggerAdapter which applies a prefix including the `id(self)` -- the
advantage of that is that every message logged by a client is
identifiable by the generating client. This may be useful in
multithreaded or multiprocess contexts, but even in single-threaded
applications, multiple clients of the same class can make things
confusing. Harmless noise when there is only one client at a time.