Tweak default env detection to fix logging
Because this function was being invoked at the time of class definition
to set defaults on the client initializers, importing these classes
before starting logging configuration effectively suppressed logging for
the environment load. That's highly undesirable: log output should
clearly indicate when we are running against a non-prod env. To resolve,
simply defer execution of that default load until the initializer is
already running (and use a sentinel value of None).
Also do some minor cleanup on the BaseClient subclasses to rely more on
passing all keyword args through to the parent.