Update TransferClient to accept UUID for endpoints
Anywhere that we were assuming something to be an endpoint_id as a
string, we now merely assume that it has a `__str__` which produces an
endpoint ID. This captures any cases where the SDK is given a uuid.UUID,
is fully backwards compatible, and supports potential future types. For
example, you could have your own `GlobusEndpoint` type, and as long as
its `__str__` is the endpoint ID, it's safe to pass to SDK methods.

Implemented with a "safe_stringify" helper that utf-8 decodes
bytestrings to be python3 compatible.