Bugfix on ClientLogAdapter
On logging.Logger objects, `warn()` is an alias for the `warning()`
method, so it's common to write `logger.warn(...)`. It's terse and
equally unambiguous.
However, LoggerAdapter objects don't have this method (at least in
py2.7, possibly in later pythons too). As a result, anything which
triggers a client warning, like a deprecated method usage, will
crashfail in py2.7