Ignore W503 and W504
From what we've been able to glean, since pycodestyle only lists W503 right
now, there is a potential change coming here. W503 enforces that binary
ops must precede newlines, W504 enforces that they must be preceded by
newlines (in multiline conditionals). This is a pretty unimportant rule
in our view, so, to avoid a potential mess in the future when this
changes, if it does, we're disabling both.