Update PaginatedResource to support MARKER paging
The Task Successful Transfers API uses a different style of paging from
other resources. Supporting this required some refactoring of the
PaginatedResource implementation.
In particular, we need to ask the PaginatedResource to count the results
it returns and to hold the marker as an instance attribute in order to
handle the fact that we can't rely on the information from offset and
limit to count things out for us.
Tried to make this refactor also productive / clarifying in terms of how
Paging counts results &c. (i.e. documented the things that I
re-learned).