Remove unsupported things from config docs
The configuration doc was really a page to note that we allow a slightly
permissive form of INI.
But we don't do that anymore.
It was also documenting the token config values.
But we don't have those anymore.

Stripped this down to a stub. I guess we can keep it for now, in case we
have a non-internal use for it in the future.