Allow num_results=None to mean unlimited results
On paginated TransferClient methods, num_results=None will be
"unlimited" results. However, this needs to be capped by the API limit.
As a result, it no seems coherent to raise an exception if too many
results are requested -- the number should just be capped to the API
limit.

As a piece of fallout from that, we no longer need
PaginationOverrunError. Removed from docs and marked as deprecated.