Relocated testsuite to prepare for new tests
The testsuite should not be inside of the package because it risks
accidentally packaging up the test files and shipping them. We remove
this risk by having tests as a top-level package, at the cost of relying
on `tests` being a discoverable package in the path. As long as testing
is being done via `make test`, this is a non-issue.

Also updated CONTRIBUTING guide to say that changes must pass
`make test`