Switch away from env var, put tokens in repo
Loading from an env var is not actually more secure, and gives us issues
with pull requests and local builds. Switch to tokens in the test
framework's set of constants, get rid of the TokenCollection, and load
the constant directly via import.

Fix minor bugs with the TransferClient.get_endpoint test:
- ensure that test_config_loading does proper teardown
- add __contains__ to GlobusResponse, letting unittest use `assertIn`