Add example of 3-legged OAuth
This is the minimal example that I could construe, using Flask to
provide the routing and redirection capabilities necessary.
The example will be all but unintelligible without some basic
understanding of Flask or similar tools, but without that, it's not
really feasible to write an example.

Not ideal, but much better than no example at all.