Remove support for access tokens from config
The access tokens loaded directly out of config are a feature we added
in the beginning of SDK development when we didn't have good solutions
for token acquisition. However, we've wanted to deprecate and remove
this functionality pretty much ever since.
This commit removes this functionality once and for all.