Remove doc references to clients sans authorizers
Any instance of "TransferClient()" or "AuthClient()" implies that these
client objects can be successfully constructed without authorizers.
However, doing so will result in clients that can't actually talk to the
services, so these docs are misleading.
Replaces the empty args with "(...)", which implies arguments but does
not provide them. The examples aren't copy-paste-able, but that's not
really their purpose anyway.