Speed up builds on Appveyor; add pypy to Travis
To reduce the inflated number of versions we cover on Appveyor, remove a
number of the 32-bit builds. I don't think we care very much about that
element of the build matrix.
Specifically, remove all of the Python3 on 32-bit system builds. We'll
keep the 2.7 on 32-bit build so that we hit a 32 bit machine at some
point. py3 on a non-64-bit box is pretty unusual.

In another direction, it would be nice to have Travis capture pypy
status, even if it's in the `allow_failures` mode.

In both cases, parallel builds are a premium feature, so we're not
going to be able to build in parallel using CI features. We could do
parallel builds on a single executor, maybe, in the future.
Additionally, simplify the Appveyor config per direction given by
packaging.python.org ; requires the addition of build.cmd