Remove confusing top-level attrs of Token Response
OAuthTokenResponse top-level properties for accessing the top-level
token aren't the desired/canonical/safe way to access tokens. We're
telling people to always use `by_resource_server`, so we should cut down
the interface for this class to that and only that.

Minor change included: make the dependent token response a subclass of
the typical token response, only overriding a helper in __init__. Also
explicitly raises NotImplementedError if someone tries to decode an ID
token with it.

Closes #158