Update travis and Makefile to use [jwt] extra
Removes the 'optional-dependencies' file, since this is now specified
via the [jwt] extra. Unfortunately, setuptools doesn't have great
support for getting extras installed on its own, and for _some reason_
`pip install .[jwt]` won't do the trick. Seems that doing the editable
install is the only way to get it done.

Dropped the `test/no-opts` make target. I dobut anyone was using it,
it's just a convenience, and it's hard to implement against the
extras_require.