Fix the broken contract with on_refresh
The RenewingAuthorizer broke our contract with supplied `on_refresh`
callbacks. Those are supposed to be called on the raw response -- to do
whatever manipulation they would like. Instead, they were being called
on the `by_resource_server` dict property.

Fix this by separating the steps of getting a token (and response) from
processing it. This might not be the cleanest imaginable fix, but it
works.