Improve Auth flow docs on 3-legged flows
Native App and base 3-legged OAuth flow params are now directly
documented on the oauth2_start_flow() methods, and that documentation is
somewhat improved. Example linkage into the examples section, and
external documentation linkage has been added there as well.

This results in duplicated docs -- for both the oauth2_start_flow()
helpers, and the underlying class initializers. If that's the cost of
having better docs, so be it.