Crank up task_wait() limit and period
Puts a couple of ints into the test constants so that we can adjust the
"default" task_wait() behavior in the future without a slew of
file-by-file edits.

Many of our tests run with a timeout of 30s and a polling interval of
1s, but there's no guarantee that tests will complete within that period
of time. Rather than trying to make the tests truly robust to an
arbitrary number of concurrent runs, it's simpler to tune things up for
now. If want more robustness in the future, we'll need to find a way of
running the testsuite under arbitrary users, and letting people run it
as their own test accounts.

The default behavior for task_wait() is now to wait a full 2 minutes and
poll on 3 a second interval. This means that builds could take a
longer time, but in practice they probably won't. Worst case test run
times increase by 6x90s = 540s = 9 minutes per environment in the build
matrix, but we'll rarely see such bad behavior.