Add example to docs of retrying task submission
Add a case to the `advanced_transfer` examples, in which we define a
fairly generic retry handler, with some healthy logging, and handle
`NetworkError, GlobusAPIError` well. Shows an application of that
handler to the TransferData structure submitted via
TransferClient.submit_transfer. The whole example aims for minimal
exposition.
It will read comfortably to anyone familiar with the idea of defining
functions on-the-fly, but very novice developers will likely struggle
with it (which is why it's advanced).