Retry NetworkErrors in testsuite
It's quite common to get spurious failures in the testsuite because one
GET somewhere along the line times out. These should all be retried, but
we don't want to put painstaking retry logic into each testcase.
Add a decorator which sets up retry behavior to the tests.framework
package, and decorate every network-sensitive test with it. It is
possible to write a test such that the retry will fail, potentially
worse than it would have failed without the retry, but (1) that's not
the common case and (2) you can use `retry_errors` on a locally defined
function to restrict its behavior if you're writing a sensitive test.
Although this is pretty much universally applied, it should be a
per-test decision, so keep the decorator pattern very simple to shuffle
around and re-apply (at the cost of some redundancy).