Add a (beta) SearchClient class for Globus Search
This client class is an adapter for the Globus Search API. The docs are
still coming online, so rather than embedding links, just lead off with
a link to the current docs.
Includes some modifications to common parts of the SDK: add the search
URL to library config, and reorder things in top-level `__init__.py` to
make a bit more sense.

Note that although there are iterable result types involved, there is no
specialized response class yet. As such, you need to key into results to
iterate. In part, this is because Search does not always place iterable
items under the same key, so some of the simplifying assumptions
applicable to Transfer don't apply here.
Fancy response objects might be significantly more effort (and perhaps
less desirable) in this case, though paginated results are likely going
to need to be handled.