Add the SearchQuery helper object
This is similar to TransferData for TransferClient. It's a specialized
class of dict with some helper methods for manipulating the contents of
the dict. In the case of SearchQuery, since we're trying to inject
multiple distinct types of data into the dict, these make sense as a set
of chaining helper methods which all return `self`. It leads to nice
usage of the form `x = q.a().b().c()`