Fix response tests to fit new interface
Now that resposne instantiation requires that a client be passed in,
all of the tests which instantiate test responses need to fit that
pattern. In many cases, the client object is never used, so we can just
pass None. In the OAuthTokenResponse tests, we should pass the mock
client object (and start testing new usage).