Remove support and docs for optional dependencies
Remove the opt dep error type, simplify usage of `jwt` to not try-except
on its import, remove the opt deps docs page from the main TOC, but
leave a stub version of the page in place. That way, anyone seeing a
link to this page from the old exception message won't be sent to a 404
page right away. We can drop it at some point.

Remove OptionalDependency stuff from tests.

Closes #258