First draft of `by_scopes` attr on token response
This is a dict-like object with a fancy __getitem__, __contains__, and
__iter__. Technically it's not a full Mapping type because it doesn't
implement the full Mapping interface. However, the only expected usage
is __contains__ and __getitem__, so this should be fine.
Even __iter__ may be an unnecessary bell on it, but maybe it will be
useful for some use cases.

Add 'resource_server' as a value in the token data we build for each
resource server. The reason is simply that this is compatible with
current by_resource_server data, and makes it a lot simpler to implement
by_scopes.

The `by_scopes` attr is actually a special object which acts as a
"scopes getter". It implements these "fancy" behaviors.
Includes tests which explore some of the basic functionality not only
against normal token response objects, but also dependent token
responses.

Resolves #183