Be robust to errors on connection establishment
If requests fails/takes too long to connect to a remote host, it raises
a specialized exception type which is safe to retry (ConnectTimeout).
Add a very simple 5x retries with 0.5s sleep to the SDK, all hardcoded,
to try to endure these failures gracefully.
All other requests exceptions are wrapped in GlobusConnectionError as
before, as is a 5th ConnectTimeout error.