GlobusConnectionTimeoutError for ConnectTimeout
When requests connections timeout, they are safe to retry (regardless of
the method being used) because no data was ever sent. In order to let
SDK clients write safe code with retries, we need to expose that error
info in some way. GlobusConnectionTimeoutError wraps
requests.ConnectTimeout for exactly this purpose.