Add connection timeout to BaseClient
Comes with a default value from config, so you can change the default
easily with an env var, and can be set explicitly when a client is
instantiated.

Gets used by clients on calls to `self._session.request()`

Documented as a `float` value, but this can also be a tuple of floats,
as accepted by `requests`. We won't make any effort to disallow that,
but don't document/explain it.

Default of 60s is high enough for all reasonable cases.