Remove local GCP delattr endpoint_id from examples
The basic usage examples, targetted at new users, should not include the
deleter for endpoint_id. That's "advanced usage", which we don't need to
explain to people right off the bat. They can always instantiate a new
object to grab an updated endpoint ID as well.

Still includes very brief mention of `del`, so that it's minimally
documented.