Add GlobusSDKUsageError & replace most ValueErrors
The GlobusSDKUsageError is a GlobusError which is also a ValueError. It
replaces many uses of ValueError to indicate that an SDK method has been
misused. This acts as an additional hint to users that they haven't read
docs carefully enough, and makes us more conformant to our stated policy
of only raising GlobusErrors unless otherwise noted.

Include new exception in docs, and tweak text to mention its existence.

Minor fix included: sphinx was warning and skipping on one of the flow
manager objects -- fix the text on that page (which hopefully nobody is
really reading and using...) to link correctly.

Resolves #139