Make clients safe(r) to un/pickle
A response should be safe to pickle and unpickle, thus guaranteeing
that we have some ready mode of support for simple caching of response
objects. Users can, of course, add attributes and break this
themselves, but that's not supported usage.

Pickling responses is currently often unsafe, as the client object
referenced by a response object has a logger attached to it with open
file handles (which themselves cannot be pickled, obviously).

To resolve, implement __getstate__ and __setstate__ which scrub and
recreate the LoggerAdapter used by client objects. This isn't quite the
same as preserving that object -- it could have handlers and other data
directly attached to it which will be lost/reinitialized -- but it is
close.

So long as the logger isn't being directly modified by a user, this
will continue to work. If it is modified, there's not much we can do.

Closes #280