Fix minor error in testsuite
This test checks that using the wrong client type for dependent tokens
gets back a 401. That *was* the behavior of Auth, but it was not spec
compliant, so it changed in a recent release.
The spec says to return 400, so check for that instead.

I'm not 100% convinced that we should be testing this, but it's
easier to adjust the test and move on than to think about
removing/replacing it.