Fix handling of `environment="production"`
When passed explicitly, this value bypassed the check we do while
looking up the environment variable, and then results in failed config
lookups.
Force the given `environment` through that config lookup code which
translates `production` to `default` and logs the use of non-production
environments.

Minor note: the logging used to say `GLOBUS_SDK_ENVIRONMENT=...` with a
strong implication that it matched the env var. Since this is no longer
always taken from there, I've modified it to just say
`globus_environment=...`

closes #306