Make RenewingAuthorizer expiration check public
Simply rename:
-    def _check_expiration_time(self):
+    def check_expiration_time(self):
so that it's official.

Also improve documentation around Renewing Authorizer and its
subclasses. In order to do this without writing some complex sphinx
extension code or other shenanigans, I accepted a bit of repetition. It
makes the docs a tiny bit harder to maintain in exchange for making the
docs a lot easier to maintain.