Add response pickleability tests
Required rewriting the mock response object to be a custom type of mock
which has a spec but resets `__class__` to be an explicit value.
Without this, pickling doesn't work and the tests can't complete.
Has to define __setstate__, __getstate__, and __reduce__ in order to
work with all of the pickling protocols.

The actual tests are on the simple side. Just pickle and unpickle
responses and ensure that they still have working properties.